Added `SpotifyPanel`
